# docs/swagger.yaml
basePath: /
definitions:
...
  refund.DecisionBody:
    properties:
      note:
        example: Reembolso autorizado
        type: string
    type: object
  refund.EligibilityResponse:
    properties:
      allowed:
        example: true
        type: boolean
      amountCents:
        description: |-
          AmountCents is what would come back to the BUYER and FeeCents how much of
          it is our commission. Both are pointers and both are OMITTED for an
          organiser: see SeesPlatformShare in usecases/refund.

          Omitted rather than zeroed. A zero here would render as "R$ 0,00 comes
          back", which is worse than saying nothing.
        example: 11000
        type: integer
      feeCents:
        example: 1000
        type: integer
      organiserCents:
        description: |-
          OrganiserCents is the part of the refund that comes out of the
          organiser's revenue. Always sent, to everyone.
        example: 10000
        type: integer
      refundsFees:
        description: |-
          RefundsFees says the fee is included. Rendered beside the amount because
          "R$ 110,00, taxa incluída" is the sentence that prevents the dispute.
        example: true
        type: boolean
      refusal:
        description: |-
          Refusal is the stable name of the reason, for a client to branch on:
          window_closed, too_close_to_event, not_paid, already_refunded,
          event_passed, request_open.
        example: too_close_to_event
        type: string
      request:
        allOf:
        - $ref: '#/definitions/refund.Response'
        description: Request is the one already in flight, when there is one.
      until:
        description: |-
          Until is when the right to cancel expires, so the page can render a real
          date instead of "soon". Absent for grounds that have no deadline.
        type: string
    type: object
  refund.Envelope:
    properties:
      data:
        $ref: '#/definitions/refund.Response'
    type: object
  refund.ErrorResponse:
    properties:
      code:
        type: string
      error:
        type: string
    type: object
  refund.ListEnvelope:
    properties:
      data:
        items:
          $ref: '#/definitions/refund.Response'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  refund.RequestBody:
    properties:
      note:
        example: Não vou conseguir ir
        type: string
      reason:
        description: |-
          Reason is the grounds. A buyer may send buyer_withdrawal or
          event_cancelled and nothing else; the other two are refused for them,
          because they are the grounds with no window.
        enum:
        - buyer_withdrawal
        - event_cancelled
        - organiser_goodwill
        example: buyer_withdrawal
        type: string
    type: object
  refund.Response:
    properties:
      amountCents:
        description: |-
          AmountCents is what goes back to the BUYER and FeeCents how much of it is
          our commission. Omitted for an organiser; see SeesPlatformShare.
        example: 11000
        type: integer
      autoApproved:
        description: |-
          AutoApproved is true when the POLICY decided rather than a person, which
          is what the screen shows instead of naming a reviewer.
        example: true
        type: boolean
      createdAt:
        type: string
      decidedAt:
        type: string
      decidedBy:
        type: string
      decisionNote:
        type: string
      eventId:
        example: evt_88
        type: string
      feeCents:
        example: 1000
        type: integer
      id:
        example: rfr_9f2c1d8a
        type: string
      note:
        type: string
      orderId:
        example: ord_4a1b7c
        type: string
      organiserCents:
        description: OrganiserCents is what comes out of the organiser's revenue.
          Always sent.
        example: 10000
        type: integer
      reason:
        example: buyer_withdrawal
        type: string
      requestedBy:
        example: usr_31
        type: string
      status:
        description: |-
          Status is pending, approved or rejected: where the DECISION stands.

          Whether the money has actually arrived is a different question and is
          answered by `orderStatus` below, which is the order's own. Two fields
          rather than one combined status, because only one of them is stored and
          the two therefore cannot disagree; a single "completed" written here
          would be a second writer for a fact the settlement path already owns.
        example: approved
        type: string
      updatedAt:
        type: string
    type: object
...
info:
  contact: {}
  description: API base para autenticação e gestão de tickets, organizada com Clean
    Architecture.
  title: Vozkot Tickets API
  version: "1.0"
paths:
...
  /api/v1/orders/{id}/refund-eligibility:
    get:
      description: 'Responde se o comprador ainda pode cancelar e receber o reembolso,
        até quando, e quanto receberia de volta. Quando não pode, devolve o motivo:
        `window_closed` (passaram os 7 dias do arrependimento, CDC art. 49), `too_close_to_event`
        (faltam menos de 48h para o evento), `not_paid`, `already_refunded`, `event_passed`
        ou `request_open`. É a mesma função que o endpoint de solicitação usa para
        decidir, então a tela e a regra nunca divergem.'
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/refund.EligibilityResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Consultar se um pedido pode ser cancelado
      tags:
      - Reembolsos
  /api/v1/orders/{id}/refund-request:
    post:
      consumes:
      - application/json
      description: 'Abre um pedido de reembolso. Dentro do prazo legal de arrependimento
        (7 dias da compra e no mínimo 48h antes do evento) e em caso de evento cancelado,
        o pedido é APROVADO na hora e o estorno entra na fila: são hipóteses que o
        organizador não tem como recusar. Fora disso, o pedido fica pendente para
        o organizador decidir. O valor devolvido inclui a taxa de serviço, conforme
        entendimento do Procon-SP e do STJ. Pedir duas vezes devolve 409.'
      parameters:
      - description: ID do pedido
        in: path
        name: id
        required: true
        type: string
      - description: Motivo
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/refund.RequestBody'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/refund.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "403":
          description: O motivo informado não pode ser alegado por este usuário
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "409":
          description: Já existe um pedido de reembolso em andamento
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "422":
          description: Fora do prazo, ou pedido não reembolsável
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Solicitar reembolso de um pedido
      tags:
      - Reembolsos
...
  /api/v1/refund-requests:
    get:
      description: A caixa de entrada do organizador e o histórico do comprador. Sem
        `eventId`, lista os pedidos do próprio comprador; com `eventId`, lista os
        do evento, e exige ser o organizador dele. Pendentes primeiro.
      parameters:
      - description: Pedidos de reembolso de um evento (exige ser o organizador)
        in: query
        name: eventId
        type: string
      - description: Filtrar por situação
        enum:
        - pending
        - approved
        - rejected
        in: query
        name: status
        type: string
      - description: Somente os que ainda ocupam o pedido (pendentes e aprovados)
        in: query
        name: open
        type: boolean
      - description: Itens por página (padrão 20, máximo 100)
        in: query
        name: limit
        type: integer
      - description: Deslocamento da paginação
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/refund.ListEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Listar pedidos de reembolso
      tags:
      - Reembolsos
  /api/v1/refund-requests/{id}:
    get:
      parameters:
      - description: ID do pedido de reembolso
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/refund.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Consultar um pedido de reembolso
      tags:
      - Reembolsos
  /api/v1/refund-requests/{id}/approve:
    post:
      consumes:
      - application/json
      description: 'Autoriza o estorno e o coloca na fila durável, na mesma transação:
        uma aprovação sem estorno, ou um estorno sem aprovação registrada, são estados
        que não podem existir. Restrito ao organizador do evento e a administradores.
        Aprovar duas vezes devolve 409, porque a segunda é outra pessoa discordando
        da primeira.'
      parameters:
      - description: ID do pedido de reembolso
        in: path
        name: id
        required: true
        type: string
      - description: Observação da decisão
        in: body
        name: request
        schema:
          $ref: '#/definitions/refund.DecisionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/refund.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "409":
          description: Já decidido
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Aprovar um pedido de reembolso
      tags:
      - Reembolsos
  /api/v1/refund-requests/{id}/reject:
    post:
      consumes:
      - application/json
      description: 'Recusa o pedido, com observação. A recusa é registrada e não altera
        o pedido de compra: é o caso sobre o qual o suporte será perguntado depois,
        então fica guardado. Restrito ao organizador do evento e a administradores.'
      parameters:
      - description: ID do pedido de reembolso
        in: path
        name: id
        required: true
        type: string
      - description: Observação da decisão
        in: body
        name: request
        schema:
          $ref: '#/definitions/refund.DecisionBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/refund.Envelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
        "409":
          description: Já decidido
          schema:
            $ref: '#/definitions/refund.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Recusar um pedido de reembolso
      tags:
      - Reembolsos
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Use o formato: Bearer {token}. Navegadores também podem autenticar
      por cookies httpOnly.'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"